@@ -1,11 +1,11 @@
 name: double_back_to_close_app
 description: A Flutter package that allows Android users to press the back-button twice to close the app.
-version: 1.2.0
+version: 1.3.0-nullsafety.1
 homepage: https://github.com/hcbpassos/double_back_to_close_app
 issue_tracker: https://github.com/hcbpassos/double_back_to_close_app/issues
 
 environment:
-  sdk: ">=2.0.0 <3.0.0"
+  sdk: '>=2.12.0-0 <3.0.0'
 
 dependencies:
   flutter:
