@@ -21,12 +21,10 @@ class DoubleBackToCloseApp extends StatefulWidget {
   /// Creates a widget that allows the user to close the app by double tapping
   /// the back-button.
   const DoubleBackToCloseApp({
-    Key key,
-    @required this.snackBar,
-    @required this.child,
-  })  : assert(snackBar != null),
-        assert(child != null),
-        super(key: key);
+    Key? key,
+    required this.snackBar,
+    required this.child,
+  }) : super(key: key);
 
   @override
   _DoubleBackToCloseAppState createState() => _DoubleBackToCloseAppState();
@@ -34,7 +32,7 @@ class DoubleBackToCloseApp extends StatefulWidget {
 
 class _DoubleBackToCloseAppState extends State<DoubleBackToCloseApp> {
   /// The last time the user tapped Android's back-button.
-  DateTime _lastTimeBackButtonWasTapped;
+  DateTime? _lastTimeBackButtonWasTapped;
 
   /// Returns whether the current platform is Android.
   bool get _isAndroid => Theme.of(context).platform == TargetPlatform.android;
@@ -47,10 +45,12 @@ class _DoubleBackToCloseAppState extends State<DoubleBackToCloseApp> {
   ///
   /// This is not quite accurate since the snack-bar could've been dismissed by
   /// the user, so this algorithm needs to be improved, as described in #2.
-  bool get _isSnackBarVisible =>
-      (_lastTimeBackButtonWasTapped != null) &&
-      (widget.snackBar.duration >
-          DateTime.now().difference(_lastTimeBackButtonWasTapped));
+  bool get _isSnackBarVisible {
+    final lastTimeBackButtonWasTapped = _lastTimeBackButtonWasTapped;
+    return (lastTimeBackButtonWasTapped != null) &&
+        (widget.snackBar.duration >
+            DateTime.now().difference(lastTimeBackButtonWasTapped));
+  }
 
   /// Returns whether the next back navigation of this route will be handled
   /// internally.
@@ -59,7 +59,7 @@ class _DoubleBackToCloseAppState extends State<DoubleBackToCloseApp> {
   /// local-history of the current route, in order to handle pop. This is done
   /// by [Drawer], for example, so it can close on pop.
   bool get _willHandlePopInternally =>
-      ModalRoute.of(context).willHandlePopInternally;
+      ModalRoute.of(context)?.willHandlePopInternally ?? false;
 
   @override
   Widget build(BuildContext context) {
@@ -81,15 +81,16 @@ class _DoubleBackToCloseAppState extends State<DoubleBackToCloseApp> {
       return true;
     } else {
       _lastTimeBackButtonWasTapped = DateTime.now();
-      Scaffold.of(context).hideCurrentSnackBar();
-      Scaffold.of(context).showSnackBar(widget.snackBar);
+      final scaffoldMessenger = ScaffoldMessenger.of(context);
+      scaffoldMessenger.hideCurrentSnackBar();
+      scaffoldMessenger.showSnackBar(widget.snackBar);
       return false;
     }
   }
 
   /// Throws a [FlutterError] if this widget was not wrapped in a [Scaffold].
   void _ensureThatContextContainsScaffold() {
-    if (Scaffold.of(context, nullOk: true) == null) {
+    if (Scaffold.maybeOf(context) == null) {
       throw FlutterError(
         '`DoubleBackToCloseApp` must be wrapped in a `Scaffold`.',
       );
